--- a/crates/extension_host/Cargo.toml
+++ b/crates/extension_host/Cargo.toml
@@ -31,6 +31,7 @@ http_client.workspace = true
 language.workspace = true
 log.workspace = true
 lsp.workspace = true
+moka.workspace = true
 node_runtime.workspace = true
 paths.workspace = true
 project.workspace = true