--- a/crates/extension_host/src/wasm_host.rs
+++ b/crates/extension_host/src/wasm_host.rs
@@ -27,9 +27,10 @@ use node_runtime::NodeRuntime;
 use release_channel::ReleaseChannel;
 use semantic_version::SemanticVersion;
 use std::borrow::Cow;
+use std::sync::LazyLock;
 use std::{
     path::{Path, PathBuf},
-    sync::{Arc, OnceLock},
+    sync::Arc,
 };
 use wasmtime::{
     CacheStore, Engine, Store,
@@ -411,19 +412,23 @@ type ExtensionCall = Box<
 >;
 
 fn wasm_engine() -> wasmtime::Engine {
-    static WASM_ENGINE: OnceLock<wasmtime::Engine> = OnceLock::new();
+    static WASM_ENGINE: LazyLock<wasmtime::Engine> = LazyLock::new(|| {
+        let mut config = wasmtime::Config::new();
+        config.wasm_component_model(true);
+        config.async_support(true);
+        config
+            .enable_incremental_compilation(cache_store())
+            .unwrap();
+        wasmtime::Engine::new(&config).unwrap()
+    });
 
-    WASM_ENGINE
-        .get_or_init(|| {
-            let mut config = wasmtime::Config::new();
-            config.wasm_component_model(true);
-            config.async_support(true);
-            config
-                .enable_incremental_compilation(Arc::new(IncrementalCompilationCache::new()))
-                .unwrap();
-            wasmtime::Engine::new(&config).unwrap()
-        })
-        .clone()
+    WASM_ENGINE.clone()
+}
+
+fn cache_store() -> Arc<IncrementalCompilationCache> {
+    static CACHE_STORE: LazyLock<Arc<IncrementalCompilationCache>> =
+        LazyLock::new(|| Arc::new(IncrementalCompilationCache::new()));
+    CACHE_STORE.clone()
 }
 
 impl WasmHost {
@@ -683,9 +688,11 @@ struct IncrementalCompilationCache {
 impl IncrementalCompilationCache {
     fn new() -> Self {
         let cache = Cache::builder()
-            // Cap this at 32 MB for now
+            // Cap this at 32 MB for now. Our extensions turn into roughly 512kb in the cache,
+            // which means we could store 64 completely novel extensions in the cache, but in
+            // practice we will more than that, which is more than enough for our use case.
             .max_capacity(32 * 1024 * 1024)
-            .weigher(|_, v: &Vec<u8>| v.len().try_into().unwrap_or(u32::MAX))
+            .weigher(|k: &Vec<u8>, v: &Vec<u8>| (k.len() + v.len()).try_into().unwrap_or(u32::MAX))
             .build();
         Self { cache }
     }
@@ -699,5 +706,101 @@ impl CacheStore for IncrementalCompilationCache {
     fn insert(&self, key: &[u8], value: Vec<u8>) -> bool {
         self.cache.insert(key.to_vec(), value);
         true
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+
+    use extension::{
+        ExtensionCapability, ExtensionLibraryKind, LanguageServerManifestEntry, LibManifestEntry,
+        SchemaVersion,
+        extension_builder::{CompileExtensionOptions, ExtensionBuilder},
+    };
+    use gpui::TestAppContext;
+    use reqwest_client::ReqwestClient;
+
+    use super::*;
+
+    #[gpui::test]
+    fn test_cache_size_for_test_extension(cx: &TestAppContext) {
+        let cache_store = cache_store();
+        let engine = wasm_engine();
+        let wasm_bytes = wasm_bytes(cx, &mut manifest());
+
+        Component::new(&engine, wasm_bytes).unwrap();
+
+        cache_store.cache.run_pending_tasks();
+        let size: usize = cache_store
+            .cache
+            .iter()
+            .map(|(k, v)| k.len() + v.len())
+            .sum();
+        // If this assertion fails, it means extensions got larger and we may want to
+        // reconsider our cache size.
+        assert!(size < 512 * 1024);
+    }
+
+    fn wasm_bytes(cx: &TestAppContext, manifest: &mut ExtensionManifest) -> Vec<u8> {
+        let extension_builder = extension_builder();
+        let path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+            .parent()
+            .unwrap()
+            .parent()
+            .unwrap()
+            .join("extensions/test-extension");
+        cx.executor()
+            .block(extension_builder.compile_extension(
+                &path,
+                manifest,
+                CompileExtensionOptions { release: true },
+            ))
+            .unwrap();
+        std::fs::read(path.join("extension.wasm")).unwrap()
+    }
+
+    fn extension_builder() -> ExtensionBuilder {
+        let user_agent = format!(
+            "Zed Extension CLI/{} ({}; {})",
+            env!("CARGO_PKG_VERSION"),
+            std::env::consts::OS,
+            std::env::consts::ARCH
+        );
+        let http_client = Arc::new(ReqwestClient::user_agent(&user_agent).unwrap());
+        // Local dir so that we don't have to download it on every run
+        let build_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("benches/.build");
+        ExtensionBuilder::new(http_client, build_dir)
+    }
+
+    fn manifest() -> ExtensionManifest {
+        ExtensionManifest {
+            id: "test-extension".into(),
+            name: "Test Extension".into(),
+            version: "0.1.0".into(),
+            schema_version: SchemaVersion(1),
+            description: Some("An extension for use in tests.".into()),
+            authors: Vec::new(),
+            repository: None,
+            themes: Default::default(),
+            icon_themes: Vec::new(),
+            lib: LibManifestEntry {
+                kind: Some(ExtensionLibraryKind::Rust),
+                version: Some(SemanticVersion::new(0, 1, 0)),
+            },
+            languages: Vec::new(),
+            grammars: BTreeMap::default(),
+            language_servers: [("gleam".into(), LanguageServerManifestEntry::default())]
+                .into_iter()
+                .collect(),
+            context_servers: BTreeMap::default(),
+            slash_commands: BTreeMap::default(),
+            indexed_docs_providers: BTreeMap::default(),
+            snippets: None,
+            capabilities: vec![ExtensionCapability::ProcessExec {
+                command: "echo".into(),
+                args: vec!["hello!".into()],
+            }],
+        }
     }
 }